Attach SDL_Window to CameraWindowSize rather than use InputSystem
Listening for a window size event with the InputSystem assumes there is only one window. To adjust the system to work with multiple windows, I would have to check the window pointer in the event data. To do that I would have to attach the window to the CameraWindowSize. But if CameraWindowSize has a handle to the window, there's no need to use events because the size could be polled from the handle.

+++ Catastrophe/Game/camera.cpp
@@ -40,7 +40,7 @@ void Camera::update(const float delta) {
   const float zoomTarget = zoomTrack.calcZoomTarget(props);
   const float newPPM = zoom.calcPPM(props, zoomTarget, delta);
   
-  const glm::vec2 newWindowSize = windowSize.get();
+  const glm::ivec2 newWindowSize = windowSize.get();
   
   props.center = newCenter;
   props.pixelsPerMeter = newPPM;
